+++ code/unix/unix_main.c
@@ -708,11 +708,11 @@ changed the load procedure to match VFS logic, and allow developer use
 extern char   *FS_BuildOSPath( const char *base, const char *game, const char *qpath );
 
 void *Sys_LoadDll( const char *name, char *fqpath ,
-                   int (**entryPoint)(int, ...),
-                   int (*systemcalls)(int, ...) ) 
+                   long (**entryPoint)(long, ...),
+                   long (*systemcalls)(long, ...) ) 
 {
   void *libHandle;
-  void  (*dllEntry)( int (*syscallptr)(int, ...) );
+  void  (*dllEntry)( long (*syscallptr)(long, ...) );
   char  curpath[MAX_OSPATH];
   char  fname[MAX_OSPATH];
   char  *basepath;
